refactor(schemas): migrate Phase 14 endpoint schemas — 3 files, 15 classes (#6042)
- analytics_evolution.py: migrate QualitySnapshot→EvolutionQualitySnapshot,
  PatternSnapshot, EvolutionTimelineRequest (dead-code), EvolutionAnalysisRequest,
  EvolutionAnalysisResponse → schemas_analytics.py
- analytics_conversation.py: migrate IntentPattern, ConversationFlow,
  ConversationMetrics, FlowBottleneck, ConversationAnalysisResult → schemas_analytics.py
- knowledge_grounding.py: migrate GroundResponseRequest, VerifyClaimRequest,
  ResolveConflictRequest, GroundedResponseSchema, ConflictSchema → schemas_knowledge.py

Renames to avoid future cross-file collisions:
- QualitySnapshot → EvolutionQualitySnapshot (analytics_quality.py also has QualitySnapshot)

Cleanup: removed dead-code imports in knowledge_grounding.py (ConflictResolution,
GroundedResponse, SearchRequest, ConflictSchema, GroundedResponseSchema were
never used in endpoint logic).
