+++ autobot-backend/api/knowledge_grounding.py
@@ -23,26 +23,25 @@ Rate limiting: 50 req/min per user for ground-response, 100 req/min for verify-c
 """
 
 import logging
-from typing import Any, Dict, List, Optional
+from typing import Any, Dict, Optional
 
 from fastapi import APIRouter, Depends, HTTPException, Path, Query, Request
-from pydantic import BaseModel, Field
 
-from api.schemas_knowledge import SearchRequest
 from api.schemas_knowledge import (
+    GroundResponseRequest,
     KnowledgeConflictsListResponse,
     KnowledgeGroundResponseResponse,
     KnowledgeGroundingStatsResponse,
     KnowledgeResolveConflictResponse,
     KnowledgeVerifyClaimResponse,
+    ResolveConflictRequest,
+    VerifyClaimRequest,
 )
 from auth_middleware import check_admin_permission, get_current_user
 from autobot_shared.error_boundaries import ErrorCategory, with_error_handling
 from constants.threshold_constants import QueryDefaults
 from services.grounded_agent import (
     Claim,
-    ConflictResolution,
-    GroundedResponse,
     get_grounded_agent,
 )
 
@@ -55,61 +54,6 @@ router = APIRouter(
 
 
 # ===== REQUEST/RESPONSE MODELS =====
-
-
-class GroundResponseRequest(BaseModel):
-    """Request to ground an agent response."""
-
-    query: str = Field(..., min_length=1, max_length=2000, description="User query")
-    agent_response: str = Field(
-        ..., min_length=1, max_length=5000, description="Agent response to ground"
-    )
-    context: Optional[Dict[str, Any]] = Field(
-        None, description="Optional context metadata"
-    )
-
-
-class VerifyClaimRequest(BaseModel):
-    """Request to verify a single claim."""
-
-    claim_text: str = Field(..., min_length=1, max_length=500)
-    subject: Optional[str] = Field(None, max_length=200)
-    predicate: Optional[str] = Field(None, max_length=200)
-    object: Optional[str] = Field(None, max_length=200)
-
-
-class ResolveConflictRequest(BaseModel):
-    """Request to resolve a conflict."""
-
-    chosen_fact: str = Field(..., min_length=1, description="Chosen fact ID")
-    reasoning: str = Field(..., min_length=10, max_length=1000)
-
-
-class GroundedResponseSchema(BaseModel):
-    """Schema for grounded response API response."""
-
-    response_id: str
-    original_query: str
-    response_text: str
-    verified_claims: List[Dict[str, Any]]
-    unverified_claims: List[Dict[str, Any]]
-    conflicts: List[Dict[str, Any]]
-    confidence_overall: float
-    requires_human_review: bool
-    timestamp: float
-
-
-class ConflictSchema(BaseModel):
-    """Schema for a conflict."""
-
-    conflict_id: str
-    claim_1_id: str
-    claim_2_id: Optional[str] = None
-    description: str
-    severity: str
-    resolution: str
-    chosen_fact: Optional[str] = None
-    timestamp: float
 
 
 # ===== ENDPOINTS =====
